[spirv-ll] Move OpenCL opcode definitions to cpp file
