+++ modules/compiler/spirv-ll/source/opcodes.cpp
@@ -2145,16 +2145,6 @@ spv::Id OpAssumeTrueKHR::Condition() const { return getValueAtOffset(1); }
 spv::Id OpExpectKHR::Value() const { return getValueAtOffset(3); }
 spv::Id OpExpectKHR::ExpectedValue() const { return getValueAtOffset(4); }
 
-spv::Id OpenCLstd::Printf::format() const { return getValueAtOffset(5); }
-
-llvm::SmallVector<spv::Id, 8> OpenCLstd::Printf::AdditionalArguments() const {
-  llvm::SmallVector<spv::Id, 8> additionalArguments;
-  for (uint16_t i = 6; i < wordCount(); i++) {
-    additionalArguments.push_back(getValueAtOffset(i));
-  }
-  return additionalArguments;
-}
-
 std::string getCapabilityName(spv::Capability cap) {
   // Note: this can't be a switch because there are multiple capability names
   // with the same enum value. We must provide a full mapping from string to
